// ebk_mobile/src/app/config/enum.ts
export enum TypeContentEnum {
  videos = 'videos',
  audios = 'audios',
  livres = 'livres',
  live = 'live',
  programme = 'programme',
  images = 'images',
  annonces = 'annonces',
  communiques = 'communiques',
  eglises = 'eglises',
  forum = 'forum',
  sujetForum = 'sujetForum',
  bibleStudy = 'bibleStudy',
  bibleStudyContent = 'bibleStudyContent',
  testimonials = 'testimonials',
}

export enum typePanoramaEnum {
  New_Testament = 'New-testament',
  Old_Testament = 'Old-testament',
}

export enum PrivilegesEnum {
  FIDELE = 'fidele_eglise',
  ADMIN_EGLISE = 'administrateur_eglise',
}

export enum AppointmentEnum {
  AWAITING = 'awaiting',
  CANCEL = 'cancel',
  APPROUVED = 'approuved',
  RECEIVED = 'received',
  POSTPONE = 'postpone',
}

export enum StatusAcounteEnum {
  ACTIF = 'actif',
  INACTIF = 'inactif',
}

export enum TestimonialStatusEnum {
  APPROVED = 'approved',
  DISMISS = 'dismiss',
  PENDING = 'pending',
}

export enum ConfirmAppointmentEnum {
  AWAITING = 'awaiting',
  CANCEL = 'cancel',
  APPROUVED = 'approuved',
  RECEIVED = 'received',
  POSTPONE = 'postpone',
}

export enum SondageQuestionTypeEnum {
  TRICOLOR = 'tricolor',
  LADDER = 'ladder',
  MCO = 'multiple_choice_open',
  MCC = 'multiple_choice_close',
  MCOT = 'multiple_choice_other',
}